Fix signed/unsigned mismatch warning in RGBController_LogitechG815.cpp

+++ Controllers/LogitechController/LogitechG815Controller/RGBController_LogitechG815.cpp
@@ -333,7 +333,7 @@ void RGBController_LogitechG815::DeviceUpdateLEDs()
         zone = ( leds[led_idx].value >> 8 );
         idx  = ( leds[led_idx].value );
 
-        if (current_colors[led_idx]==new_colors[led_idx])
+        if(current_colors[led_idx]==new_colors[led_idx])
         {
             /*-------------------------------------------------*\
             | Don't send if key color is not changed            |
@@ -366,7 +366,7 @@ void RGBController_LogitechG815::DeviceUpdateLEDs()
 
         colorkey = new_colors[led_idx];
 
-        if (ledsByColors.count(colorkey) == 0)
+        if(ledsByColors.count(colorkey) == 0)
         {
             ledsByColors.insert(std::pair<RGBColor, std::vector<char>>(colorkey, {}));
         }
@@ -400,16 +400,16 @@ void RGBController_LogitechG815::DeviceUpdateLEDs()
                 frame_buffer_big_mode[2] = RGBGetBValue(x.first);
                 frame_pos                = 3;
 
-                for(uint8_t i = 0; i < max_key_per_color; i++)
+                for(uint8_t i = 0; i < (uint8_t)max_key_per_color; i++)
                 {
-                    if(bi + i < x.second.size())
+                    if((bi + i) < (uint8_t)x.second.size())
                     {
                         frame_buffer_big_mode[frame_pos] = x.second[bi+i];
                         frame_pos++;
                     }
                 }
 
-                if (frame_pos < data_size)
+                if(frame_pos < data_size)
                 {
                     /*-----------------------------------------*\
                     | Zeroing just what is needed and if needed |
@@ -446,7 +446,7 @@ void RGBController_LogitechG815::DeviceUpdateLEDs()
                 li++;
                 led_in_little_frame++;
 
-                if (led_in_little_frame == 4)
+                if(led_in_little_frame == 4)
                 {
                     /*-----------------------------------------*\
                     | No End of Data byte if the packet is full |
